Updated config scripts and views
