+++ scripts/config.js
@@ -1,0 +1,15 @@
+const rpcs = {
+  1: 'https://ethereum.publicnode.com',
+  42161: 'https://arb1.arbitrum.io/rpc',
+  56: 'https://bsc-dataseed.bnbchain.org',
+  80001: 'https://rpc.ankr.com/polygon_mumbai'
+}
+
+const store = {
+  1: '0x6579dF8f986e4A982F200DAfa0c1b955A438f620',
+  56: '0x6579dF8f986e4A982F200DAfa0c1b955A438f620',
+  42161: '0x6579dF8f986e4A982F200DAfa0c1b955A438f620',
+  80001: '0xA483aF3eD0DF092EC4b9b93DA400f84922c92Be9'
+}
+
+export { store, rpcs }
